Day03/requirements.txt
beautifulsoup4
redis
pyyaml